requirements.txt
streamlit
google-generativeai
python-dotenv
numpy
sounddevice
